Day 15

+++ src/template.rs
@@ -1,12 +1,10 @@
-use std::str;
-
-pub async fn advent(data: String) -> u32 {
+pub async fn advent(data: String) -> usize {
     let mut answer = 0;
 
     return answer;
 }
 
-pub async fn advent_2(data: String) -> u32 {
+pub async fn advent_2(data: String) -> usize {
     let mut answer = 0;
 
     return answer;
